+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+## Added
+
+- Ability to delete lists ([#58](https://github.com/garritfra/fling/pull/58))
+
 ## v0.6.3 (2023-01-03)
 
 ## Improvements